+++ src/main/java/racingcar/domain/FixedGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class FixedGenerator implements NumberGenerator{
