+++ src/main/java/racingcar/domain/MovingCondition.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class MovingCondition {
     public static final int MOVABLE_MIN_VALUE = 4;
