+++ src/main/java/racingcar/domain/RacingCarGame.java
@@ -1,6 +1,6 @@
-package racingcar;
+package racingcar.domain;
 
-import racingcar.ui.ResultView;
+import racingcar.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.Collections;
