+++ src/test/java/racingcar/MovingConditionTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.MovingCondition;
 
 import static org.assertj.core.api.Assertions.assertThat;
