+++ src/test/java/racingcar/RacingCarGameTest.java
@@ -2,7 +2,11 @@ package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.ui.ResultView;
+import racingcar.domain.Car;
+import racingcar.domain.FixedGenerator;
+import racingcar.domain.MovingCondition;
+import racingcar.domain.RacingCarGame;
+import racingcar.view.ResultView;
 
 import java.util.Arrays;
 import java.util.List;
